backend/requirements.txt
# API clients
tweepy>=4.14.0
openai>=1.3.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0